requirements.txt
Django>=4.2,<5
gunicorn
psycopg2-binary
dj-database-url
djangorestframework==3.14